Add contextMenu for (un)staged files
The menupopup elements cannot be added as the first child of the window.
If being added as the first child, the following error happens:
org.eclipse.jface.action.MenuManager cannot be cast to org.eclipse.swt.widgets.Composite
This is because GitPerspective.getUI() tries to cast the first child of the window to Composite.